+++ web/app/components/horizontalcolorbar.js
